Znalezienie zespolu

+++ muzyka/src/main/java/projekt/dao/Dao.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface Dao<T> {
-    Optional<T> get(int id);
+    Optional<T> get(Long id);
     List<T> getAll();
     void save(T o);
     void update(T o) throws IllegalArgumentException;
